# api/users/docs/swagger.yaml
definitions:
  github_com_xiaohangshuhub_admin_internal_users_app_user.CreateCmd:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/github_com_xiaohangshuhub_admin_internal_users_domain_user.Gender'
        description: 性别
      nickname:
        description: 昵称
        type: string
      phone:
        description: 手机号
        type: string
      pwd:
        description: 密码
        type: string
      roles:
        description: 角色
        items:
          type: string
        type: array
      salt:
        description: 密码盐值
        type: string
      username:
        description: 用户名
        type: string
    type: object
  github_com_xiaohangshuhub_admin_internal_users_app_user.Login:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  github_com_xiaohangshuhub_admin_internal_users_app_user.UserDto:
    properties:
      access_token:
        type: string
      email:
        type: string
      nikename:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  github_com_xiaohangshuhub_admin_internal_users_domain_user.Gender:
    enum:
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      Female: 女
      Male: 男
      Unknown: 保密
    x-enum-descriptions:
    - 保密
    - 男
    - 女
    x-enum-varnames:
    - Unknown
    - Male
    - Female
  internal_users_webapi.Response-bool:
    properties:
      code:
        type: integer
      data:
        type: boolean
      message:
        type: string
    type: object
  internal_users_webapi.Response-github_com_xiaohangshuhub_admin_internal_users_app_user_UserDto:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/github_com_xiaohangshuhub_admin_internal_users_app_user.UserDto'
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /user:
    post:
      consumes:
      - application/json
      description: 创建新用户123...
      parameters:
      - description: 创建用户参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xiaohangshuhub_admin_internal_users_app_user.CreateCmd'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_users_webapi.Response-bool'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_users_webapi.Response-bool'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_users_webapi.Response-bool'
      summary: 创建用户
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录验证
      parameters:
      - description: 登录参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xiaohangshuhub_admin_internal_users_app_user.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/internal_users_webapi.Response-github_com_xiaohangshuhub_admin_internal_users_app_user_UserDto'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_users_webapi.Response-github_com_xiaohangshuhub_admin_internal_users_app_user_UserDto'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_users_webapi.Response-github_com_xiaohangshuhub_admin_internal_users_app_user_UserDto'
      summary: 用户登录
      tags:
      - User
swagger: "2.0"